+++ controllers/Product_Controller.php
@@ -29,7 +29,7 @@ class ProductController {
         $subcategoryModel = new Subcategory_Model();
         $subcategories = $subcategoryModel->getAllSubcategories();
 
-        $page = isset($_GET['page']) ? (int)$_GET['page'] : 1;
+        $page = isset($_GET['page']) ? (int)Utils::limpiar_datos($_GET['page']) : 1;
         $items_per_page = 4; // Puedes cambiar este valor según tus necesidades
         $productos = $productModel->getProductsByPage($page, $items_per_page);
         $total_users = $productModel->getProductCount();
@@ -39,12 +39,12 @@ class ProductController {
 
     public function registerProduct() {
         if ($_SERVER['REQUEST_METHOD'] == 'POST') {
-            $nombre = $_POST['nombre'];
-            $descripcion = $_POST['descripcion'];
-            $precio = $_POST['precio'];
-            $stock = $_POST['stock'];
-            $categoria = $_POST['categoria'];
-            $subcategoria = $_POST['subcategoria'];
+            $nombre = Utils::limpiar_datos($_POST['nombre']);
+            $descripcion = Utils::limpiar_datos($_POST['descripcion']);
+            $precio = Utils::limpiar_datos($_POST['precio']);
+            $stock = Utils::limpiar_datos($_POST['stock']);
+            $categoria = Utils::limpiar_datos($_POST['categoria']);
+            $subcategoria = Utils::limpiar_datos($_POST['subcategoria']);
             $imagenes = [];
             for ($i = 1; $i <= 6; $i++) {
                 $imagen = isset($_FILES['imagen' . $i]) && $_FILES['imagen' . $i]['error'] === UPLOAD_ERR_OK ? $_FILES['imagen' . $i]['name'] : null;
@@ -73,7 +73,7 @@ class ProductController {
         if (isset($_GET['id'])) {
             $role = $_SESSION['rol'];
             $productModel = new Product_Model();
-            $producto = $productModel->getProductById($_GET['id']);
+            $producto = $productModel->getProductById(Utils::limpiar_datos($_GET['id']));
             $categoryModel = new Category_Model();
             $categories = $categoryModel->getAllCategories();
             $subcategoryModel = new Subcategory_Model();
@@ -84,13 +84,13 @@ class ProductController {
 
     public function updateProduct() {
         if ($_SERVER['REQUEST_METHOD'] == 'POST') {
-            $id = $_POST['id'];
-            $nombre = $_POST['nombre'];
-            $descripcion = $_POST['descripcion'];
-            $precio = $_POST['precio'];
-            $stock = $_POST['stock'];
-            $categoria = $_POST['categoria'];
-            $subcategoria = $_POST['subcategoria'];
+            $id = Utils::limpiar_datos($_POST['id']);
+            $nombre = Utils::limpiar_datos($_POST['nombre']);
+            $descripcion = Utils::limpiar_datos($_POST['descripcion']);
+            $precio = Utils::limpiar_datos($_POST['precio']);
+            $stock = Utils::limpiar_datos($_POST['stock']);
+            $categoria = Utils::limpiar_datos($_POST['categoria']);
+            $subcategoria = Utils::limpiar_datos($_POST['subcategoria']);
             $imagenes = [];
             for ($i = 1; $i <= 6; $i++) {
                 $imagen = isset($_FILES['imagen' . $i]) && $_FILES['imagen' . $i]['error'] === UPLOAD_ERR_OK ? $_FILES['imagen' . $i]['name'] : null;
@@ -106,24 +106,28 @@ class ProductController {
 
             $productModel = new Product_Model();
             $productModel->updateProduct($id, $nombre, $descripcion, $precio, $stock, $categoria, $subcategoria, $imagenes);
-            var_dump($productModel);
             header('Location: Product_Controller.php?action=listProducts');
         }
     }
-
     public function deleteProduct() {
         if (isset($_POST['id'])) {
-            $productId = $_POST['id'];
+            $productId = Utils::limpiar_datos($_POST['id']);
             $productModel = new Product_Model();
-            $productModel->deleteProduct($productId);
-
-            header('Location: Product_Controller.php?action=listProducts');
+            
+            // Verificar si el producto está asociado a un pedido
+            if ($productModel->isProductInOrder($productId)) {
+                echo "<script>alert('Este producto está asociado a un pedido y no puede ser eliminado.'); window.location.href='Product_Controller.php?action=listProducts';</script>";
+            } else {
+                $productModel->deleteProduct($productId);
+                echo "<script>alert('Producto eliminado correctamente.'); window.location.href='Product_Controller.php?action=listProducts';</script>";
+            }
+            exit();
         }
     }
-
+    
     public function getSubcategories() {
         if ($_SERVER['REQUEST_METHOD'] == 'POST' && isset($_POST['categoryId'])) {
-            $categoryId = $_POST['categoryId'];
+            $categoryId = Utils::limpiar_datos($_POST['categoryId']);
             $categoryModel = new Category_Model();
             $subcategories = $categoryModel->getSubcategoriesByCategoryId($categoryId);
             echo json_encode($subcategories);
@@ -137,12 +141,12 @@ class ProductController {
 
 // Manejo de acciones basadas en los parámetros de URL
 if (isset($_GET['action'])) {
-    $action = $_GET['action'];
+    $action = Utils::limpiar_datos($_GET['action']);
     $productController = new ProductController();
 
     switch ($action) {
         case 'listProducts':
-            $page = isset($_GET['page']) ? (int)$_GET['page'] : 1;
+            $page = isset($_GET['page']) ? (int)Utils::limpiar_datos($_GET['page']) : 1;
             $productController->listProducts($page);
             break;
         case 'editProduct':
