+++ controllers/User_Controller.php
@@ -20,28 +20,35 @@ class UserController {
 
     public function register_Admin() {
         if ($_SERVER['REQUEST_METHOD'] == 'POST') {
-            $nombre = $_POST['nombre'];
-            $apellido = $_POST['apellido'];
-            $direccion = $_POST['direccion'];
-            $telefono = $_POST['telefono'];
-            $email = $_POST['email'];
-    
+            $nombre = Utils::limpiar_datos($_POST['nombre']);
+            $apellido = Utils::limpiar_datos($_POST['apellido']);
+            $direccion = Utils::limpiar_datos($_POST['direccion']);
+            $telefono = Utils::limpiar_datos($_POST['telefono']);
+            $email = Utils::limpiar_datos($_POST['email']);
+
+            $userModel = new User_Model();
+
+            // Comprobar si el correo electrónico ya está registrado
+            if ($userModel->emailExists($email)) {
+                echo 'Error: El correo electrónico ya está registrado.';
+                return;
+            }
+
             // Generar contraseña aleatoria
             $password = Utils::generateRandomPassword();
             $hashedPassword = password_hash($password, PASSWORD_DEFAULT);
-    
+
             $codigo_activacion = Utils::generar_codigo_activacion();
             $imagen_perfil = isset($_FILES['imagen_perfil']) ? $_FILES['imagen_perfil']['name'] : null;
-    
+
             if ($imagen_perfil) {
                 $target_dir = __DIR__ . "/../Assets/images/uploads/";
                 $target_file = $target_dir . basename($imagen_perfil);
                 move_uploaded_file($_FILES['imagen_perfil']['tmp_name'], $target_file);
             }
-    
-            $userModel = new User_Model();
+
             $userModel->createUser($nombre, $apellido, $direccion, $telefono, $email, $hashedPassword, $codigo_activacion, $imagen_perfil);
-    
+
             if (Utils::enviarCorreoActivacionAdmin($email, $codigo_activacion, $password)) {
                 //echo 'Registro exitoso. Por favor, revisa tu correo electrónico para activar tu cuenta.';
                 header('Location: User_Controller.php?action=listUsers');
@@ -53,11 +60,9 @@ class UserController {
         }
     }
     
-    
-
     public function listUsers() {
         $role = $_SESSION['rol'];
-        $page = isset($_GET['page']) ? (int)$_GET['page'] : 1;
+        $page = isset($_GET['page']) ? (int)Utils::limpiar_datos($_GET['page']) : 1;
         $items_per_page = 10; // Puedes cambiar este valor según tus necesidades
     
         $userModel = new User_Model();
@@ -72,21 +77,21 @@ class UserController {
         if (isset($_GET['id'])) {
             $role = $_SESSION['rol'];
             $userModel = new User_Model();
-            $usuario = $userModel->getUserById($_GET['id']);
+            $usuario = $userModel->getUserById(Utils::limpiar_datos($_GET['id']));
             require_once '../views/user/edit_user_views.php';
         }
     }
 
     public function updateUser() {
         if ($_SERVER['REQUEST_METHOD'] == 'POST') {
-            $id = $_POST['id'];
-            $nombre = $_POST['nombre'];
-            $apellido = $_POST['apellido'];
-            $direccion = $_POST['direccion'];
-            $telefono = $_POST['telefono'];
-            $email = $_POST['email'];
-            $rol = $_POST['rol'];
-            $estado = $_POST['estado'];
+            $id = Utils::limpiar_datos($_POST['id']);
+            $nombre = Utils::limpiar_datos($_POST['nombre']);
+            $apellido = Utils::limpiar_datos($_POST['apellido']);
+            $direccion = Utils::limpiar_datos($_POST['direccion']);
+            $telefono = Utils::limpiar_datos($_POST['telefono']);
+            $email = Utils::limpiar_datos($_POST['email']);
+            $rol = Utils::limpiar_datos($_POST['rol']);
+            $estado = Utils::limpiar_datos($_POST['estado']);
             $imagen_perfil = isset($_FILES['imagen_perfil']) && $_FILES['imagen_perfil']['error'] === UPLOAD_ERR_OK ? $_FILES['imagen_perfil']['name'] : null;
     
             // Subir imagen de perfil si existe
@@ -103,10 +108,9 @@ class UserController {
         }
     }
     
-
     public function deleteUser() {
         if (isset($_POST['id'])) {
-            $userId = $_POST['id'];
+            $userId = Utils::limpiar_datos($_POST['id']);
             $userModel = new User_Model();
     
             // Comprobar si el usuario tiene pedidos
@@ -120,11 +124,10 @@ class UserController {
         }
     }
     
-
     public function banUser() {
         if (isset($_POST['id'])) {
-            $userId = $_POST['id'];
-            $action = $_POST['action'];
+            $userId = Utils::limpiar_datos($_POST['id']);
+            $action = Utils::limpiar_datos($_POST['action']);
 
             $userModel = new User_Model();
 
@@ -140,8 +143,8 @@ class UserController {
 
     public function activate() {
         if ($_SERVER['REQUEST_METHOD'] == 'POST' && isset($_POST['email']) && isset($_POST['codigoActivacion'])) {
-            $email = $_POST['email'];
-            $codigo = $_POST['codigoActivacion'];
+            $email = Utils::limpiar_datos($_POST['email']);
+            $codigo = Utils::limpiar_datos($_POST['codigoActivacion']);
 
             $userModel = new User_Model();
             $user = $userModel->getUserByEmailAndCode($email, $codigo);
@@ -160,12 +163,12 @@ class UserController {
 
 // Manejo de acciones basadas en los parámetros de URL
 if (isset($_GET['action'])) {
-    $action = $_GET['action'];
+    $action = Utils::limpiar_datos($_GET['action']);
     $userController = new UserController();
 
     switch ($action) {
         case 'listUsers':
-            $page = isset($_GET['page']) ? (int)$_GET['page'] : 1;
+            $page = isset($_GET['page']) ? (int)Utils::limpiar_datos($_GET['page']) : 1;
             $userController->listUsers($page);
             break;
         case 'editUser':
@@ -193,6 +196,3 @@ if (isset($_GET['action'])) {
 } else {
     //echo 'No se especificó ninguna acción';
 }
-
-
-
